updated plot functions
